dev-requirements.txt
affine==2.3.1
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
branca==0.5.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.0.5
cryptography
cycler==0.11.0
defusedxml==0.7.1
Django==4.1.1
django-allauth==0.51.0
django-bootstrap-icons==0.8.1
django-bootstrap-v5==1.0.11
django-extra-views==0.14.0
Fiona==1.8.21
flopy==3.3.5
folium==0.12.1.post1
fonttools==4.37.3
frozenlist==1.3.3
GeoAlchemy2==0.12.5
geopandas==0.11.1
greenlet==1.1.3
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.0
multidict==6.0.3
munch==2.5.0
numpy==1.23.3
oauthlib==3.2.1
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
pip==22.0.2
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
pyproj==3.4.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.2.1
rasterio==1.3.2
requests==2.28.1
requests-oauthlib==1.3.1
scipy==1.9.1
setuptools==59.6.0
Shapely==1.8.4
six==1.16.0
snuggs==1.4.7
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
sqlparse==0.4.3
tqdm==4.64.1
types-cryptography==3.3.23
Unipath==1.1
urllib3==1.26.12
vtk==9.2.2
whitenoise==6.2.0
wslink==1.9.2
yarl==1.8.2
docker==6.0.0
python-decouple
colour
pyvista
xmltodict